Widen activation code to u64
Replace the restricted-length u32 type used for account activation codes
and simply rely on u64s.  As part of this and for simplicity reasons,
persist them as i64 in the database, accepting that sometimes they will
show up as negative.

+++ authn/src/db/postgres.sql
@@ -25,7 +25,10 @@ CREATE TABLE IF NOT EXISTS users (
     email VARCHAR(64) UNIQUE NOT NULL,
 
     -- Activation code.  If present, the account has not been activated yet.
-    activation_code INT,
+    --
+    -- Note that this is supposed to be an u64 so it can show up as negative when persisted
+    -- in the database.
+    activation_code BIGINT,
 
     -- The user's last successful login timestamp.
     last_login TIMESTAMPTZ
